requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==45.0.2
dj-database-url==2.3.0
Django==5.2.1
django-cors-headers==4.3.1
django-filter==25.1
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.2
frozenlist==1.6.0
idna==3.10
kombu==5.5.3
magic-filter==1.0.12
multidict==6.4.3
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
Pillow==10.4.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==6.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.0
